feat: refactor gatewayConfig into small components
